+++ airbnb/review/src/main/java/com/example/review/domain/Review.java
@@ -1,8 +1,5 @@
-package com.example.review.model;
+package com.example.review.domain;
 
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Data;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
