+++ airbnb/review/src/main/java/com/example/review/domain/event/HostReviewCreated.java
@@ -1,9 +1,11 @@
-package com.example.review.model;
+package com.example.review.domain.event;
+
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
+
 import java.time.LocalDateTime;
 
 @AllArgsConstructor
@@ -11,7 +13,7 @@ import java.time.LocalDateTime;
 @Getter
 @Builder
 @Accessors(chain = true)
-public class CustomerReviewCreated {
+public class HostReviewCreated {
     private Long id;
 
     private String commentBy;
