+++ airbnb/review/src/main/java/com/example/review/domain/event/HostReviewUpdated.java
@@ -1,4 +1,4 @@
-package com.example.review.model;
+package com.example.review.domain.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
@@ -6,12 +6,19 @@ import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
 
+import java.time.LocalDateTime;
 
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Builder
 @Accessors(chain = true)
-public class RoomReviewDeleted {
+public class HostReviewUpdated {
     private Long id;
+
+    private String description;
+
+    private Double score;
+
+    private LocalDateTime commentDate;
 }
