+++ airbnb/review/src/main/java/com/example/review/domain/event/RoomReviewCreated.java
@@ -1,20 +1,26 @@
-package com.example.review.model;
+package com.example.review.domain.event;
 
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
 import lombok.experimental.Accessors;
+import lombok.extern.slf4j.Slf4j;
 
 import java.time.LocalDateTime;
 
+@Slf4j
 @AllArgsConstructor
 @NoArgsConstructor
 @Getter
 @Builder
 @Accessors(chain = true)
-public class HostReviewUpdated {
+public class RoomReviewCreated {
     private Long id;
+
+    private String commentBy;
+
+    private String commentTo;
 
     private String description;
