+++ airbnb/review/src/main/java/com/example/review/domain/event/RoomReviewDeleted.java
@@ -1,0 +1,17 @@
+package com.example.review.domain.event;
+
+import lombok.AllArgsConstructor;
+import lombok.Builder;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.experimental.Accessors;
+
+
+@AllArgsConstructor
+@NoArgsConstructor
+@Getter
+@Builder
+@Accessors(chain = true)
+public class RoomReviewDeleted {
+    private Long id;
+}
