+++ airbnb/review/src/main/java/com/example/review/repository/ReviewRepository.java
@@ -1,5 +1,5 @@
 package com.example.review.repository;
-import com.example.review.model.Review;
+import com.example.review.domain.Review;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
