+++ airbnb/review/src/main/java/com/example/review/support/error/ReviewError.java
@@ -7,6 +7,7 @@ import org.springframework.http.HttpStatus;
 public enum ReviewError {
     ENTITY_NOT_FOUND("Cannot find review entity", HttpStatus.NOT_FOUND),
     DUPLICATE("Duplicate input",HttpStatus.BAD_REQUEST),
+    WRONG_FORMAT("Wrong input",HttpStatus.BAD_REQUEST),
     NOT_PERMITTED("Create review not allowed",HttpStatus.FORBIDDEN);
 
     private String description;
